requirements/development.txt
asgiref==3.8.1
Django==5.1.1
django-cleanup==9.0.0
django-cors-headers==4.4.0
django-dirtyfields==1.9.3
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.7
packaging==24.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.7.0
